+++ lib/td_audit/support/query_support.ex
@@ -23,6 +23,14 @@ defmodule TdAudit.QuerySupport do
     dynamic([p], field(p, ^atom_key) > ^value and ^acc)
   end
 
+  defp filter_by_type(:start_ts, value, acc) do
+    dynamic([p], field(p, :ts) >= ^value and ^acc)
+  end
+
+  defp filter_by_type(:end_ts, value, acc) do
+    dynamic([p], field(p, :ts) <= ^value and ^acc)
+  end
+
   defp filter_by_type(atom_key, values, acc) when is_list(values) do
     dynamic([p], field(p, ^atom_key) in ^values and ^acc)
   end