+++ test/td_audit_web/controllers/event_controller_test.exs
@@ -64,20 +64,43 @@ defmodule TdAuditWeb.EventControllerTest do
 
     @tag authenticated_user: @admin_user_name
     test "lists all events filtered by range timestamp", %{conn: conn, swagger_schema: schema} do
+      insert(:event, event: "event_type_ts0", ts: "2021-02-01T00:00:00Z")
       %{id: id1} = insert(:event, event: "event_type_ts1", ts: "2021-02-02T00:00:00Z")
       %{id: id2} = insert(:event, event: "event_type_ts2", ts: "2021-02-03T00:00:00Z")
+      insert(:event, event: "event_type_ts3", ts: "2021-02-04T00:00:00Z")
 
       assert %{"data" => data} =
               conn
               |> get(
                 Routes.event_path(conn, :index,
-                ts: %{between: ["2021-02-02", "2021-02-03"]}
+                  start_ts: "2021-02-02T00:00:00Z",
+                  end_ts: "2021-02-03T00:00:00Z"
                 )
               )
               |> validate_resp_schema(schema, "EventsResponse")
               |> json_response(:ok)
 
-      assert [%{"id" => ^id1}, %{"id" => ^id2}] = data
+      assert [%{"id" => ^id2}, %{"id" => ^id1}] = data
+
+    end
+
+    @tag authenticated_user: @admin_user_name
+    test "lists all events filtered by start_ts", %{conn: conn, swagger_schema: schema} do
+      insert(:event, event: "event_type_ts0", ts: "2021-02-01T00:00:00Z")
+      insert(:event, event: "event_type_ts1", ts: "2021-02-02T00:00:00Z")
+      %{id: id} = insert(:event, event: "event_type_ts2", ts: "2021-02-03T00:00:00Z")
+
+      assert %{"data" => data} =
+              conn
+              |> get(
+                Routes.event_path(conn, :index,
+                  start_ts: "2021-02-02T12:00:00Z"
+                )
+              )
+              |> validate_resp_schema(schema, "EventsResponse")
+              |> json_response(:ok)
+
+      assert [%{"id" => ^id}] = data
 
     end
   end